Create test flash and sound 2
This code makes a sound file play and LED strip illuminate with a flashing sequence when either the "a" or "b" button is pushed on cpx.  Crikit is driving the LED'S and speaker.  At this point the audio plays and then the LED'S light.  I need to fix code so that they both happen at the same time.  Then I need a break beam to sub in for the onboard switch to activate sequence.